requirements.txt
pandas
python-box
transformers
transformers[sentencepiece]
torch
pyyaml
matplotlib
fastapi
uvicorn
nltk
boto3
sacrebleu
rouge_score
Jinja2
datasets
py7zr
tqdm
mypy-boto3-s3
-e .
